requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-filter==2.4.0
django-cors-headers==3.10.1
